feat(ui): migrate presentational components for Issue #566
Move color helpers, ActionButton, PaginationButton, LazyLoad, Header, and
PageHeader into @chronogrove/ui with subpath exports; theme keeps thin shims.
Add react-intersection-observer to @chronogrove/ui; drop it from the theme.

Bump gatsby-theme-chronogrove and @chronogrove/ui to 0.78.0. Extend theme Jest
coverage ignores for UI re-export files. Document changes in CHANGELOG.

Made-with: Cursor
