+++ packages/ui/src/action-button.spec.js
@@ -1,14 +1,11 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui'
 import { render, screen, fireEvent } from '@testing-library/react'
-import { Provider } from 'react-redux'
-import { configureStore } from '@reduxjs/toolkit'
 import { ThemeUIProvider } from 'theme-ui'
 
-import ActionButton from './action-button'
-import { BUTTON_PRIMARY_COLORS } from '../utils/colors'
+import ActionButton from './action-button.js'
+import { BUTTON_PRIMARY_COLORS } from './color-utils.js'
 
-// Mock useThemeUI for fallback tests - mutable mock like root-wrapper.spec.js
 const mockUseThemeUI = jest.fn(() => ({
   colorMode: 'default',
   theme: {
@@ -24,7 +21,6 @@ jest.mock('theme-ui', () => ({
   useThemeUI: () => mockUseThemeUI()
 }))
 
-// Mock theme (primary/primaryRgb so components use theme colors)
 const mockTheme = {
   colors: {
     primary: BUTTON_PRIMARY_COLORS.light,
@@ -38,20 +34,9 @@ const mockTheme = {
   }
 }
 
-// Mock store
-const mockStore = configureStore({
-  reducer: {
-    theme: (state = { colorMode: 'light' }) => state
-  }
-})
-
 const renderWithProviders = (component, customTheme = null) => {
   const themeToUse = customTheme ?? mockTheme
-  return render(
-    <Provider store={mockStore}>
-      <ThemeUIProvider theme={themeToUse}>{component}</ThemeUIProvider>
-    </Provider>
-  )
+  return render(<ThemeUIProvider theme={themeToUse}>{component}</ThemeUIProvider>)
 }
 
 describe('ActionButton', () => {
@@ -87,7 +72,6 @@ describe('ActionButton', () => {
 
     const button = screen.getByRole('button', { name: /primary button/i })
     expect(button).toHaveStyle({ fontWeight: 'medium' })
-    // Primary uses theme color (may be hex or CSS var in Theme UI)
     expect(button).toBeInTheDocument()
   })
 
@@ -150,7 +134,6 @@ describe('ActionButton', () => {
     renderWithProviders(<ActionButton variant='invalid'>Invalid Variant</ActionButton>)
 
     const button = screen.getByRole('button', { name: /invalid variant/i })
-    // Invalid variant falls back to primary (theme colors)
     expect(button).toBeInTheDocument()
     expect(button).toHaveStyle({ fontWeight: 'medium' })
   })
@@ -159,7 +142,6 @@ describe('ActionButton', () => {
     renderWithProviders(<ActionButton size='invalid'>Invalid Size</ActionButton>)
 
     const button = screen.getByRole('button', { name: /invalid size/i })
-    // Should use medium size since invalid size falls back to medium
     expect(button).toHaveStyle({
       fontSize: '12px',
       padding: '8px 12px'
@@ -168,7 +150,6 @@ describe('ActionButton', () => {
 
   describe('theme fallbacks', () => {
     beforeEach(() => {
-      // Reset to default mock
       mockUseThemeUI.mockReturnValue({
         colorMode: 'default',
         theme: {
@@ -181,10 +162,9 @@ describe('ActionButton', () => {
     })
 
     it('uses fallback primary color when theme.colors.primary is undefined', () => {
-      // Mock useThemeUI to return theme without primary to hit fallback branch (line 20)
       mockUseThemeUI.mockReturnValueOnce({
         colorMode: 'default',
-        theme: { colors: {} } // No primary property - triggers fallback '#422EA3'
+        theme: { colors: {} }
       })
 
       renderWithProviders(<ActionButton>Fallback Test</ActionButton>)
@@ -195,10 +175,9 @@ describe('ActionButton', () => {
     })
 
     it('uses fallback primaryRgb when theme.colors.primaryRgb is undefined', () => {
-      // Mock useThemeUI to return theme with primary but no primaryRgb (line 24)
       mockUseThemeUI.mockReturnValueOnce({
         colorMode: 'default',
-        theme: { colors: { primary: '#422EA3' } } // Has primary, no primaryRgb - triggers fallback '66, 46, 163'
+        theme: { colors: { primary: '#422EA3' } }
       })
 
       renderWithProviders(<ActionButton>Fallback RGB Test</ActionButton>)
@@ -208,10 +187,9 @@ describe('ActionButton', () => {
     })
 
     it('uses fallback when theme itself is undefined', () => {
-      // Mock useThemeUI to return undefined theme (line 20)
       mockUseThemeUI.mockReturnValueOnce({
         colorMode: 'default',
-        theme: undefined // theme is undefined - triggers fallback '#422EA3'
+        theme: undefined
       })
 
       renderWithProviders(<ActionButton>Undefined Theme Test</ActionButton>)
