+++ packages/ui/src/pagination-button.spec.js
@@ -1,14 +1,11 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui'
 import { render, screen, fireEvent } from '@testing-library/react'
-import { Provider } from 'react-redux'
-import { configureStore } from '@reduxjs/toolkit'
 import { ThemeUIProvider } from 'theme-ui'
 
-import PaginationButton from './pagination-button'
-import { BUTTON_PRIMARY_COLORS } from '../utils/colors'
+import PaginationButton from './pagination-button.js'
+import { BUTTON_PRIMARY_COLORS } from './color-utils.js'
 
-// Mock useThemeUI for fallback tests - mutable mock like root-wrapper.spec.js
 const mockUseThemeUI = jest.fn(() => ({
   colorMode: 'default',
   theme: {
@@ -24,7 +21,6 @@ jest.mock('theme-ui', () => ({
   useThemeUI: () => mockUseThemeUI()
 }))
 
-// Mock theme (primary/primaryRgb so components use theme colors)
 const mockTheme = {
   colors: {
     primary: BUTTON_PRIMARY_COLORS.light,
@@ -38,20 +34,9 @@ const mockTheme = {
   }
 }
 
-// Mock store
-const mockStore = configureStore({
-  reducer: {
-    theme: (state = { colorMode: 'light' }) => state
-  }
-})
-
 const renderWithProviders = (component, customTheme = null) => {
   const themeToUse = customTheme ?? mockTheme
-  return render(
-    <Provider store={mockStore}>
-      <ThemeUIProvider theme={themeToUse}>{component}</ThemeUIProvider>
-    </Provider>
-  )
+  return render(<ThemeUIProvider theme={themeToUse}>{component}</ThemeUIProvider>)
 }
 
 describe('PaginationButton', () => {
@@ -99,7 +84,6 @@ describe('PaginationButton', () => {
 
     const button = screen.getByRole('button', { name: /1/i })
     expect(button).toHaveStyle({ fontWeight: 'medium' })
-    // Primary uses theme color (may be hex or CSS var in Theme UI)
     expect(button).toBeInTheDocument()
   })
 
@@ -162,7 +146,6 @@ describe('PaginationButton', () => {
 
   describe('theme fallbacks', () => {
     beforeEach(() => {
-      // Reset to default mock
       mockUseThemeUI.mockReturnValue({
         colorMode: 'default',
         theme: {
@@ -175,10 +158,9 @@ describe('PaginationButton', () => {
     })
 
     it('uses fallback primary color when theme.colors.primary is undefined', () => {
-      // Mock useThemeUI to return theme without primary to hit fallback branch (line 28)
       mockUseThemeUI.mockReturnValueOnce({
         colorMode: 'default',
-        theme: { colors: {} } // No primary property - triggers fallback '#422EA3'
+        theme: { colors: {} }
       })
 
       renderWithProviders(<PaginationButton>1</PaginationButton>)
@@ -189,10 +171,9 @@ describe('PaginationButton', () => {
     })
 
     it('uses fallback primaryRgb when theme.colors.primaryRgb is undefined', () => {
-      // Mock useThemeUI to return theme with primary but no primaryRgb (line 32)
       mockUseThemeUI.mockReturnValueOnce({
         colorMode: 'default',
-        theme: { colors: { primary: '#422EA3' } } // Has primary, no primaryRgb - triggers fallback '66, 46, 163'
+        theme: { colors: { primary: '#422EA3' } }
       })
 
       renderWithProviders(<PaginationButton>1</PaginationButton>)
@@ -202,10 +183,9 @@ describe('PaginationButton', () => {
     })
 
     it('uses fallback when theme itself is undefined', () => {
-      // Mock useThemeUI to return undefined theme (line 28)
       mockUseThemeUI.mockReturnValueOnce({
         colorMode: 'default',
-        theme: undefined // theme is undefined - triggers fallback '#422EA3'
+        theme: undefined
       })
 
       renderWithProviders(<PaginationButton>1</PaginationButton>)
