+++ theme/package.json
@@ -1,7 +1,7 @@
 {
   "name": "gatsby-theme-chronogrove",
   "description": "A personal website and digital garden theme for GatsbyJS.",
-  "version": "0.77.0",
+  "version": "0.78.0",
   "author": "Chris Vogt <mail@chrisvogt.me> (https://www.chrisvogt.me)",
   "main": "index.js",
   "license": "MIT",
@@ -108,7 +108,6 @@
     "react": "catalog:",
     "react-dom": "catalog:",
     "react-error-boundary": "^6.1.1",
-    "react-intersection-observer": "catalog:",
     "react-photo-gallery": "catalog:",
     "react-placeholder": "^4.1.0",
     "react-redux": "^9.2.0",
